backend/requirements.txt
Django
django-cors-headers
django-debug-toolbar
django-environ
django-filter
djangorestframework
djangorestframework-simplejwt
djoser
python-dotenv
psycopg2
gunicorn
